Rename 2 files with wrong file names
