+++ lib/stdio/encoding-final.js
@@ -14,11 +14,11 @@ export const handleStreamsEncoding = (stdioStreams, {encoding}, isSync) => {
 	const stringDecoder = new StringDecoder(encoding);
 	const generator = objectMode
 		? {
-			transform: encodingEndObjectGenerator.bind(undefined, stringDecoder),
+			transform: encodingObjectGenerator.bind(undefined, stringDecoder),
 		}
 		: {
-			transform: encodingEndStringGenerator.bind(undefined, stringDecoder),
-			final: encodingEndStringFinal.bind(undefined, stringDecoder),
+			transform: encodingStringGenerator.bind(undefined, stringDecoder),
+			final: encodingStringFinal.bind(undefined, stringDecoder),
 		};
 	return [
 		...newStdioStreams,
@@ -39,17 +39,17 @@ const shouldEncodeOutput = (stdioStreams, encoding, isSync) => stdioStreams[0].d
 // eslint-disable-next-line unicorn/text-encoding-identifier-case
 const IGNORED_ENCODINGS = new Set(['utf8', 'utf-8', 'buffer']);
 
-const encodingEndStringGenerator = function * (stringDecoder, chunk) {
+const encodingStringGenerator = function * (stringDecoder, chunk) {
 	yield stringDecoder.write(chunk);
 };
 
-const encodingEndStringFinal = function * (stringDecoder) {
+const encodingStringFinal = function * (stringDecoder) {
 	const lastChunk = stringDecoder.end();
 	if (lastChunk !== '') {
 		yield lastChunk;
 	}
 };
 
-const encodingEndObjectGenerator = function * (stringDecoder, chunk) {
+const encodingObjectGenerator = function * (stringDecoder, chunk) {
 	yield isUint8Array(chunk) ? stringDecoder.end(chunk) : chunk;
 };
