+++ lib/stdio/encoding-transform.js
@@ -13,19 +13,19 @@ However, those are converted to Buffer:
 - on writes: `Duplex.writable` `decodeStrings: true` default option
 - on reads: `Duplex.readable` `readableEncoding: null` default option
 */
-export const getEncodingStartGenerator = encoding => {
+export const getEncodingTransformGenerator = encoding => {
 	if (encoding === 'buffer') {
-		return {transform: encodingStartBufferGenerator.bind(undefined, new TextEncoder())};
+		return {transform: encodingBufferGenerator.bind(undefined, new TextEncoder())};
 	}
 
 	const textDecoder = new TextDecoder();
 	return {
-		transform: encodingStartStringGenerator.bind(undefined, textDecoder),
-		final: encodingStartStringFinal.bind(undefined, textDecoder),
+		transform: encodingStringGenerator.bind(undefined, textDecoder),
+		final: encodingStringFinal.bind(undefined, textDecoder),
 	};
 };
 
-const encodingStartBufferGenerator = function * (textEncoder, chunk) {
+const encodingBufferGenerator = function * (textEncoder, chunk) {
 	if (Buffer.isBuffer(chunk)) {
 		yield new Uint8Array(chunk);
 	} else if (typeof chunk === 'string') {
@@ -35,13 +35,13 @@ const encodingStartBufferGenerator = function * (textEncoder, chunk) {
 	}
 };
 
-const encodingStartStringGenerator = function * (textDecoder, chunk) {
+const encodingStringGenerator = function * (textDecoder, chunk) {
 	yield Buffer.isBuffer(chunk) || isUint8Array(chunk)
 		? textDecoder.decode(chunk, {stream: true})
 		: chunk;
 };
 
-const encodingStartStringFinal = function * (textDecoder) {
+const encodingStringFinal = function * (textDecoder) {
 	const lastChunk = textDecoder.decode();
 	if (lastChunk !== '') {
 		yield lastChunk;
