+++ lib/stdio/generator.js
@@ -1,5 +1,5 @@
 import {generatorsToTransform} from './transform.js';
-import {getEncodingStartGenerator} from './encoding-end.js';
+import {getEncodingTransformGenerator} from './encoding-transform.js';
 import {getLinesGenerator} from './lines.js';
 import {pipeStreams} from './pipeline.js';
 import {isGeneratorOptions, isAsyncGenerator} from './type.js';
@@ -73,7 +73,7 @@ export const generatorToDuplexStream = ({
 	optionName,
 }) => {
 	const generators = [
-		getEncodingStartGenerator(encoding),
+		getEncodingTransformGenerator(encoding),
 		getLinesGenerator(encoding, binary),
 		{transform, final},
 		{transform: getValidateTransformReturn(readableObjectMode, optionName)},
