+++ lib/stdio/handle.js
@@ -4,7 +4,7 @@ import {normalizeStdio} from './option.js';
 import {handleNativeStream} from './native.js';
 import {handleInputOptions} from './input.js';
 import {handleStreamsLines} from './lines.js';
-import {handleStreamsEncoding} from './encoding-start.js';
+import {handleStreamsEncoding} from './encoding-final.js';
 import {normalizeGenerators} from './generator.js';
 import {forwardStdio} from './forward.js';
