+++ test/stdio/encoding-final.js
