+++ test/stdio/encoding-transform.js
